// types/enums/Category.ts
export enum Category {
    Music = "music",
    SportAndLeisure = "sport_and_leisure",
    FilmAndTV = "film_and_tv",
    ArtsAndLiterature = "arts_and_literature",
    History = "history",
    SocietyAndCulture = "society_and_culture",
    Science = "science",
    Geography = "geography",
    FoodAndDrink = "food_and_drink",
    GeneralKnowledge = "general_knowledge"
}